backend_v2/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
email-validator
